+++ anarchive/package.json
@@ -40,12 +40,9 @@
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "pm2": "^4.4.0",
-    "react-p5": "^1.3.2",
-    "react-p5-wrapper": "^2.0.0",
     "react-three-fiber": "^4.2.10",
     "rsyncwrapper": "^3.0.1",
-    "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.0"
+    "socket.io": "^2.3.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.2"
