--- client/public/peep/index.html
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title></title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.8.0/p5.js"></script>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.8.0/addons/p5.dom.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.2.0/socket.io.js"></script>
-    <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.0.0/addons/p5.sound.js"> </script>
-  </head>
-  <body>
-    <script>
-    var socket;
-    let mySynth;
-
-
-      function setup(){
-        createCanvas(500,500);
-        background('blue');
-
-        socket = io.connect('http://localhost:3000');
-        socket.on('mouse', newDrawing);
-
-        getAudioContext().suspend();
-
-        mySynth = new p5.MonoSynth();
-         // This won't play until the context has resumed
-         mySynth.play('D6');
-
-      }
-
-      function newDrawing(data){
-        noStroke();
-        fill(255, 0, 100);
-        ellipse(data.x, data.y, 36, 36);
-        mySynth.play(data.soundNote, data.soundVel);
-      }
-
-      function mouseDragged(){
-        console.log(mouseX+ ',' + mouseY);
-        let soundNote = random(['Fb4', 'G4']);
-         // note velocity (volume, from 0 to 1)
-        let soundVel = random();
-
-        var data={
-        x: mouseX,
-        y: mouseY,
-        soundNote: soundNote,
-        soundVel: soundVel
-      }
-      // mySynth.play(data.soundNote, data.soundVel);
-
-      socket.emit('mouse', data);
-
-      fill('green');
-      ellipse(mouseX, mouseY, 36, 26);
-    }
-
-    function mousePressed() {
-      userStartAudio();
-    }
-
-      function draw(){
-
-      }
-    </script>
-
-  </body>
-</html>
